+++ release_notes.md
@@ -2,6 +2,7 @@
     * Added token-based participant account generation
     * Added support for multiple tokens per participant
     * Added support for downloading participant mapping 
+    * Improved admin for Researcher and LDAP to Researcher converter
     * Fixed some issues with development logging
 * 1.7.3
     * Fixed issues where crashes would sometimes occur in Admin portal during password change & logout